requirements.txt
alembic==0.7.5.post2
Flask==0.10.1
Flask-HTTPAuth==2.4.0
Flask-Migrate==1.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
py==1.4.26
pytest==2.7.0
pytest-flask==0.7.5
python-dateutil==2.4.2
six==1.9.0
SQLAlchemy==0.9.9
Werkzeug==0.10.1
